refactor: factor out watcher and subscription
